Remove watchdog into scripts folder.
To install without having watchdog prior to install
